refactor: split loaded objfile into focused modules
